OCD-1932: re-refactor surv bg job; csv and xml daily files broken
